+++ Plugins/ActorPools/Source/NexusActorPoolsEditor/Private/Visualizers/NActorPoolSpawnerComponentVisualizer.cpp
@@ -2,6 +2,7 @@
 // See the LICENSE file at the repository root for more information.
 
 #include "Visualizers/NActorPoolSpawnerComponentVisualizer.h"
+
 #include "SceneManagement.h"
 #include "NActorPoolSpawnerComponent.h"
 #include "NEditorUserSettings.h"
@@ -17,9 +18,6 @@ void FNActorPoolSpawnerComponentVisualizer::DrawVisualization(const UActorCompon
 	const FVector Origin = SpawnerComponent->GetComponentLocation() + SpawnerComponent->GetOffset();
 	const FRotator Rotator = SpawnerComponent->GetComponentRotation();
 	const UNEditorUserSettings* Settings = UNEditorUserSettings::Get();
-
-	constexpr FColor DistributionOuterColor = FColor(200,0,0);
-	constexpr FColor DistributionInnerColor = FColor(100,0,0);
 	
 	switch (SpawnerComponent->GetDistribution())
 	{
@@ -27,7 +25,7 @@ void FNActorPoolSpawnerComponentVisualizer::DrawVisualization(const UActorCompon
 		{
 			DrawCircle(PDI, Origin,
 				FRotationMatrix(Rotator).GetScaledAxis(EAxis::X), FRotationMatrix(Rotator).GetScaledAxis(EAxis::Y),
-				DistributionOuterColor, 5.f, Settings->DebugCircleSides, SDPG_Foreground, Settings->DebugLineThickness);
+				Settings->DistributionOuterColor, 5.f, Settings->DebugCircleSides, SDPG_Foreground, Settings->DebugLineThickness);
 			break;
 		}
 		
@@ -36,10 +34,10 @@ void FNActorPoolSpawnerComponentVisualizer::DrawVisualization(const UActorCompon
 			const FVector RadiusRange = SpawnerComponent->GetDistributionRange();
 			DrawCircle(PDI, Origin,
 				FRotationMatrix(Rotator).GetScaledAxis(EAxis::X), FRotationMatrix(Rotator).GetScaledAxis(EAxis::Y),
-				DistributionInnerColor, RadiusRange.X, Settings->DebugCircleSides, SDPG_Foreground, Settings->DebugLineThickness);
+				Settings->DistributionInnerColor, RadiusRange.X, Settings->DebugCircleSides, SDPG_Foreground, Settings->DebugLineThickness);
 			DrawCircle(PDI, Origin,
 				FRotationMatrix(Rotator).GetScaledAxis(EAxis::X), FRotationMatrix(Rotator).GetScaledAxis(EAxis::Y),
-				DistributionOuterColor, RadiusRange.Y, Settings->DebugCircleSides, SDPG_Foreground, Settings->DebugLineThickness);
+				Settings->DistributionOuterColor, RadiusRange.Y, Settings->DebugCircleSides, SDPG_Foreground, Settings->DebugLineThickness);
 			break;
 		}
 
@@ -47,8 +45,8 @@ void FNActorPoolSpawnerComponentVisualizer::DrawVisualization(const UActorCompon
 		{
 			const FVector SphereRange = SpawnerComponent->GetDistributionRange();
 			const FTransform& SphereTransform = SpawnerComponent->GetComponentTransform();
-			DrawWireSphere(PDI, SphereTransform, DistributionInnerColor, SphereRange.X, Settings->DebugCircleSides, SDPG_Foreground, Settings->DebugLineThickness);
-			DrawWireSphere(PDI, SphereTransform, DistributionOuterColor, SphereRange.Y, Settings->DebugCircleSides, SDPG_Foreground, Settings->DebugLineThickness);
+			DrawWireSphere(PDI, SphereTransform, Settings->DistributionInnerColor, SphereRange.X, Settings->DebugCircleSides, SDPG_Foreground, Settings->DebugLineThickness);
+			DrawWireSphere(PDI, SphereTransform, Settings->DistributionOuterColor, SphereRange.Y, Settings->DebugCircleSides, SDPG_Foreground, Settings->DebugLineThickness);
 			break;
 		}
 
@@ -56,7 +54,7 @@ void FNActorPoolSpawnerComponentVisualizer::DrawVisualization(const UActorCompon
 		{
 			const FVector BoxRange = SpawnerComponent->GetDistributionRange();
 			const FBox Box = FBox(Origin - BoxRange, Origin + BoxRange);
-			DrawWireBox(PDI, Box, DistributionOuterColor, SDPG_Foreground, Settings->DebugLineThickness);
+			DrawWireBox(PDI, Box, Settings->DistributionOuterColor, SDPG_Foreground, Settings->DebugLineThickness);
 			break;
 		}